Serious rework - compiles (cleanly) on Win32 with Visual Studio Express 2013.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat5@4461 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libFM/Frame.cpp
@@ -37,7 +37,7 @@ int Frame::lookupAddressForName(const Object &name, bool searchGlobals) {
 	  auto mfunc = _ctxt->_module->ro(_module)->m_locals.find(name);
 	  if (mfunc != _ctxt->_module->ro(_module)->m_locals.end())
 	    {
-	      ndx = _sym_names.count();
+	      ndx = int(_sym_names.count());
 	      _ctxt->_list->push(_sym_names,name);
 	      _ctxt->_list->push(_vars,mfunc->second);
 	      return ndx;
@@ -49,7 +49,7 @@ int Frame::lookupAddressForName(const Object &name, bool searchGlobals) {
 	{
 	  // We are a proxy for someone who wants this symbol (not us!)
 	  // Allocate a space for it
-	  ndx = _sym_names.count();
+	  ndx = int(_sym_names.count());
 	  _ctxt->_list->push(_sym_names,name);
 	  _ctxt->_list->push(_vars,_ctxt->_globals->get(fname,_ctxt));
 	  // We don't need to cache the address for this symbol
@@ -105,7 +105,7 @@ int Frame::allocateVariable(const FMString &name)
 {
   int p = getAddress(name);
   if (p != -1) return p;
-  p = _sym_names.count();
+  p = int(_sym_names.count());
   _ctxt->_list->push(_sym_names,_ctxt->_string->makeString(name));
   _ctxt->_list->push(_vars,_ctxt->_double->empty());
   return p;
@@ -113,7 +113,7 @@ int Frame::allocateVariable(const FMString &name)
 
 int Frame::defineNewSymbol(const Object &name)
 {
-  int p = _sym_names.count();
+  int p = int(_sym_names.count());
   _ctxt->_list->push(_sym_names,name);
   _ctxt->_list->push(_vars,_ctxt->_double->empty());
   return p;
@@ -154,9 +154,9 @@ int Frame::mapIPToLineNumber(int ip)
     const CodeData *dp = _ctxt->_code->ro(_code);
     if (dp->m_lineno.isEmpty()) return -1;
     rle_decode_line_nos(_ctxt->_uint32->ro(dp->m_lineno),
-			dp->m_lineno.count(),
+			unsigned(dp->m_lineno.count()),
 			_debug_line_nos);
   }
-  if (ip >= _debug_line_nos.size()) return -1;
+  if (ip >= int(_debug_line_nos.size())) return -1;
   return _debug_line_nos[ip];  
 }
